TODO: Update items for 'damo status' deprecation
We decided not to record DAMON status.  The intention was to record the
changes to kdamonds including parameter changes and CPU usages.

But, it makes no much sense to change the parameters while recording is
ongoing.  Maybe a reasonable way to deal with such case is, stopping
record and start again.  We may need to add a more graceful handling in
future (e.g., 'damo tune' check if 'damo record' is ongoing, and give
some warnings, or do the handling on its own).

For CPU usages recording, we already doing it via ProcStat.  As the
ProcStat is for not only kdamond but general processes, it is more
flexible.  For now, it records stats for kdamonds and monitoring
targets.  In future, we could extend that for damo itself.

Anyway, for the reason, we decided not to record DAMON status.  Instead,
'record' saves the given kdamonds once at the beginning.  And let
proc_stat recording does the record of the CPU utilization changes.

Updating TODO to reflect the decision.

Signed-off-by: SeongJae Park <sj@kernel.org>