_damon,damo_stop: add initialization exception for stop command
If feature_supports file is removed or corrupted while DAMON is running,
damo cannot stop it since reading it is failed, and writing it again is
prohibited due to the running DAMON.  Add an exception for stop command.
When initialization is requested by 'damo stop', simply ignore
feature_supports read failure.

Signed-off-by: SeongJae Park <sj@kernel.org>