damo_args_damon: support 'report' format
'damo args damon' output of json and yaml are human-readasble, but still
not that easy to read[1].  Add another format named 'report' for 'damo
args damon --format'.  It shows the DAMON parameters in 'damo report
damon' format.  It is for human eyes, not for machines.  Unlike
json/yaml format, the output cannot be passed to other 'damo' commands.

[1] https://github.com/damonitor/damo/issues/10

Signed-off-by: SeongJae Park <sj@kernel.org>