damo_args_damon: implement --remove option
As '--add' option for 'damo args damon' is implmented, add '--remove'
option.  It works in very similar way.  It receives file path and index
of the kdamond to remove.  Then the command open the file, load the
kdamonds parameters from it, remove the kdamond of the given index, and
print it.  Users can redirect the output to another json or yaml file.

Signed-off-by: SeongJae Park <sj@kernel.org>