damo_report_access: add an experimental option for cache usage monitoring
DAMON monitoring results can show rough cache space usage.  Support such
use case by adding an option for translating DAMON monitoring results to
mapped on a virtual cache.  The option receives size, associativity
(ways), and line size of the virtual cache.  Then, 'damo report access'
transforms the monitoring results to be mapped on the cache and show the
access pattern on the user-requested format.

The feature is experimental.  It can be changed and removed in near
future.  Also note that it doesn't support 'age' information.

Signed-off-by: SeongJae Park <sj@kernel.org>