_damon/updte_tuned_intervals(): return no error if the feature is not supported
On kernels that not supporting intervals auto-tuning, the function
fails.  Since most of the users assume it might be due to a race
condition and retries with back-off sleep, it causes hang.  Fix it by
treating error due to feature unsupport as no error.  The intervals will
anyway have the up to date value, so indeed it shouldn't be considered
as an error.

Fixes: 94cfd25d9bc6 ("damo_report_damon: show auto-tuned intervals")
Signed-off-by: SeongJae Park <sj@kernel.org>