release_note: fix wrong version numbers
I mistakenly tagged a version after v2.7.3 as v2.7.5.  Meanwhile, I
wrote release_note without knowing that.  So v2.7.4 section on the
release_note is for v2.7.5, and v2.7.5 section on the release_note is
for v2.7.6.  As v2.7.5 and v2.7.6 are already tagged and pushed, fix the
release_note.

Signed-off-by: SeongJae Park <sj@kernel.org>