damo_report_access: implement a function for inferring aggregation interval
If a record is made with auto-tuned monitoring intervals, record's
interval is the initial one, so cannot be applied to every snapshot.
Each snapshot actually has their start/end time, so we can infer the
aggregation interval for each snapshot.  Actually the snapshot-based
aggregation interval should be more accurate.  Use record's aggregation
interval if the error is not big, though, to make the logic and
resulting outputs clear and consistent.

Signed-off-by: SeongJae Park <sj@kernel.org>