_damo_records.add_childs_target: Revert to old target list on failure
_damo_records.add_childs_target goes through the list of targets, checks
if they have children, and adds those children to a list of new targets.
Those new targets get added to the kdamonds data structure's list of
targets, and then the targets are committed to DAMON. However, if the
commit fails, the new targets will be placed in kdamonds' data
structure, even though they are not tracked by DAMON. This prevents
those new processes from being tracked by DAMON until another child task
is created.

To avoid this problem, this patch reverts kdamonds' target list to the
original on failure.

Signed-off-by: Bijan Tabatabai <bijantabatab@micron.com>