_damo_records: Recursively add child tasks in add_childs_target
_damo_records.add_childs_target() currently iterates through each target,
gets a list of their child tasks, then adds those child tasks (not already
in the list of targets) to the list of targets. This function currently
only goes one level deep. If a target's child has child tasks itself, it
will take two calls to add_childs_target() to add the grandchild to the
list of targets.

This patch modifies the behavior of add_childs_target() to recursively
search for a target's (great^n)-grandchildren rather than just its
children.

Signed-off-by: Bijan Tabatabai <btabatabai@wisc.edu>