_damon: Fix lost targets in add_childs_target
Since commit 9f5063d, add_childs_target (then poll_add_childs) filters
terminated processes out of the list of new processes to add. It does
this by setting the new_targets list to include all the previously
tracked processes that are still running and then appends a child_pid to
that list. The problem is, it does this for each child_pid, meaning that
any previous child_pids that were added to the new_targets list will be
discarded.

This patch refactors add_childs_target so that all found child pids are
added to the list of targets.

Fixes: 9f5063d
Signed-off-by: Bijan Tabatabai <btabatabai@wisc.edu>