_damo_records: allow retrieving snapshot from damon_stat
If damon_stat is running and it is the only source of the information we
can get, and if the request is not specific to not use it, using it as a
last resort can be useful.  Implement a basic code for deciding whether
to use damon_stat or not.  Real snapshot retrieval is to be implemented.

Signed-off-by: SeongJae Park <sj@kernel.org>