_damo_records: implement max seconds per output record file feature
If recording is continued for a long time, the resulting record file can
be too huge to be handled in practically short time.  Add a
RcordingHandl field for setting max time length per record file.  If a
recording is continued longer than the time limit, every record
infomration that collected so far is saved under a directory of name
'<given output file name>.<timestamp>', as the file names in normal
case.  <timestamp> is in '%Y-%m-%d-%H-%M-%S' format.

Signed-off-by: SeongJae Park <sj@kernel.org>