_damo_records: set max_seconds_per_file as 3600
The feature would be most always useful.  Enable it with a default
value, one hour.  Because it is too clear to be useful, don't add a
command line for disabling it or even changing the value for now.  Maybe
adding a knob for disabling or changing the value could be useful, but
not necessary for now.

Closes: https://github.com/damonitor/damo/issues/38
Signed-off-by: SeongJae Park <sj@kernel.org>