damo_adjust: support record filters
When a too huge record file is created, it is useful to split it into
small files, for a time intervals or address ranges.  Add the record
filters arguments to 'damo adjust', so that it can be used for the
purpose.  The filters arguments are same to those for 'damo report
access'.

Closes: https://github.com/damonitor/damo/issues/37
Signed-off-by: SeongJae Park <sj@kernel.org>