_damon/add_childs_target(): do nothing for target regions fixed ops
add_childs_target() is designed for vaddr, which runs for entire virtual
address spaces of given processes.  Hence it is doing nothing if the
given context is configured for virtual address spaces.  But there is
fvaddr, which is also for virtual address spaces, but for user-specified
fixed address ranges.  add_childs_target() sets no target region since
it is a signal for vaddr to dynamically find whole mapped address ranges
and monitor those.  It is a weird signal for fvaddr, to just do no
monitoring.  Do notihng for fvaddr, too.

Signed-off-by: SeongJae Park <sj@kernel.org>