_damon/add_vaddr_child_targets_with_obsolete_support(): pass string to pid_running()
pid_running() expects to receive the pid as a string, but
add_vaddr_child_targets_with_obsolete_support() passes DamonTarget.pid,
which could be an int, as is.  As a result of this, 'damo record' can be
finished earlier than expected.  Convert the argument into a string
before passing, like add_vaddr_child_targets() does.

Fixes: c86b5aa36683 ("_damon/add_vaddr_child_targets(): manage monitorign resutls correctly on obsolete_target supporting kernel")
Signed-off-by: SeongJae Park <sj@kernel.org>